pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "organvm-schema-definitions"
version = "0.1.0"
description = "Canonical JSON Schema definitions for the organvm eight-organ system"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "jsonschema>=4.20",
    "pyyaml>=6.0",
]

[project.scripts]
organvm-validate = "scripts.validate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
